refactor(tauri/ui): use generic store helpers for all stores
