+++ tauri-app/src/lib/stores/orderDetail.ts
@@ -1,32 +1,7 @@
-import { get, writable } from 'svelte/store';
+import { get } from 'svelte/store';
 import type { Order } from '$lib/typeshare/orderDetail';
 import { invoke } from '@tauri-apps/api';
 import { subgraphUrl } from '$lib/stores/settings';
+import { useDetailStore } from '$lib/storesGeneric/detailStore';
 
-function useOrderDetailStore() {
-  const STORAGE_KEY = "orders.orderDetail";
-
-  const { subscribe, update } = writable<{[id: string]: Order}>(localStorage.getItem(STORAGE_KEY) ? JSON.parse(localStorage.getItem(STORAGE_KEY) as string) : {});
-
-  subscribe(value => {
-    if(value) {
-      localStorage.setItem(STORAGE_KEY, JSON.stringify(value));
-    } else {
-      localStorage.setItem(STORAGE_KEY, JSON.stringify({}));
-    }
-  });
-
-  async function refetch(id: string) {
-    const res: Order = await invoke("order_detail", {id, subgraphArgs: { url: get(subgraphUrl)} });
-    update((value) => {
-      return {... value, [id]: res};
-    });
-  }
-
-  return {
-    subscribe,
-    refetch
-  }
-}
-
-export const orderDetail = useOrderDetailStore();
+export const orderDetail = useDetailStore<Order>("orders.orderDetail", (id: string) => invoke("order_detail", {id, subgraphArgs: { url: get(subgraphUrl)} }));
