+++ tauri-app/src/lib/stores/ordersList.ts
@@ -2,7 +2,7 @@ import { get } from 'svelte/store';
 import type { Order } from '$lib/typeshare/ordersList';
 import { invoke } from '@tauri-apps/api';
 import { subgraphUrl } from '$lib/stores/settings';
-import { usePaginatedCachedStore } from '$lib/stores/paginatedStore';
+import { usePaginatedCachedStore } from '$lib/storesGeneric/paginatedStore';
 
 export const ordersList = usePaginatedCachedStore<Order>(
   'ordersList',
