+++ tauri-app/src/lib/stores/vaultListBalanceChanges.ts
@@ -1,10 +1,8 @@
 import { get } from 'svelte/store';
 import { invoke } from '@tauri-apps/api';
 import { subgraphUrl } from '$lib/stores/settings';
-import { usePaginatedCachedStore } from './paginatedStore';
+import { usePaginatedCachedStore } from '$lib/storesGeneric/paginatedStore';
 import type { VaultBalanceChange } from '$lib/typeshare/vaultListBalanceChanges';
-
-
 export const useVaultListBalanceChanges = (id: string) =>  usePaginatedCachedStore<VaultBalanceChange>(
   `vaultListBalanceChanges-${id}`,
   (page) => invoke("vault_list_balance_changes", {subgraphArgs: { url: get(subgraphUrl)}, id, paginationArgs: { page, page_size: 10 } }),
